Logic - Fixed regeneration and counter rules in Explore Mode - Altered Difficulty system - Fixed Annihilation logic - Fixed StackingRule for dead units (1/3 of girth now)
Misc
- Added combatLog specials

AI
- Changed Restoration mode priority calc function
- Updated AtomicAi: clearshot checks, smarter approach/turn/reload decision alg (WIP)

Visuals
- Added a hard-fix for UnitView grid-cell anchoring (in case move-anim breaks)
- Fixed tooltip "unknown"

Animations
- Added a check for visible unitView (no anims otherwise; not sure if it's bug-free yet)
- Fixed some issues with 3d anims
- Disabled auto-search

+++ Dueling-Club/src/main/java/main/libgdx/anims/AnimMultiplicator.java
@@ -244,7 +244,7 @@ public class AnimMultiplicator implements Runnable {
     private void createAndAddEmitterActions(EmitterActor actor, Coordinates c) {
 //        MoveToAction action = ActorMaster.getMoveToAction(c, actor, pixelsPerSecond);
         Vector2 v = getCenteredPos(c);
-        int speed = getPixelsPerSecond();
+        float speed = getPixelsPerSecond();
         if (template != null) {
             GridMaster.offset(
                     getOrigin(),
@@ -255,7 +255,7 @@ public class AnimMultiplicator implements Runnable {
             speed = template.speed;
         }
         MoveByAction action =
-                ActorMaster.getMoveByAction(getOrigin(), v, actor, speed);
+                ActorMaster.getMoveByAction(getOrigin(), v, actor, (int) speed);
 
 
         if (action.getDuration() > this.duration) {
@@ -316,7 +316,7 @@ public class AnimMultiplicator implements Runnable {
         return anim.getRef();
     }
 
-    public int getPixelsPerSecond() {
+    public float getPixelsPerSecond() {
         return
                 anim.getPixelsPerSecond();
     }
