+++ Dueling-Club/src/main/java/main/libgdx/anims/std/RangedAttackAnim.java
@@ -102,7 +102,7 @@ public class RangedAttackAnim extends AttackAnim {
 
 
     @Override
-    public int getPixelsPerSecond() {
+    public float getPixelsPerSecond() {
         return 750;
     }
